server/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
CodeConvert==3.0.2
Django==1.11.5
django-cors-headers==2.1.0
django-models-ext==1.1.9
django-url-shortening==1.0.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
importlib-metadata==0.23
kombu==3.0.37
more-itertools==7.2.0
pkg-resources==0.0.0
psycopg2==2.7.3.1
PyJWT==1.5.3
python-dateutil==2.8.1
pytz==2017.2
redis==2.10.3
short-url==1.2.2
shortuuid==0.5.0
six==1.13.0
TimeConvert==1.5.0
tzlocal==2.0.0
vine==1.3.0
zipp==0.6.0
